oneplus5_defconfig: Enable CONFIG_WQ_POWER_EFFICIENT_DEFAULT
Commit cee22a15052f ("workqueues: Introduce new flag WQ_POWER_EFFICIENT
for power oriented workqueues") introduced the concept of power
efficient workqueues, but it was never enabled in upstream kernel
configs.

Power efficient workqueues are simply marked as "unbound", so that jobs
queued to them can run on any CPU in the system. It leaves the target
CPU selection to the scheduler, which is the best place for such
decision making. This improves power efficiency for workqueues which are
otherwise pinned to a CPU.

Enable it for oneplus5 as ARM platforms were the main target for the
introduction of power efficient workqueues.

Signed-off-by: Michele Beccalossi <beccalossi.michele@gmail.com>