oneplus5_defconfig: Set default readahead to 128KB
A lower readahead value, specifically 128KB, proved to give a
performance benefit to the runtime execution, as opposed to boot time
benefitting from a higher value. Since the boot times are minimally
affected, lower this value for better runtime performance.

It would be better to keep a high readahead value while booting and then
lowering it after the boot process is done; this would require
additional changes to the Android init process I want to avoid for the
moment.

Signed-off-by: Michele Beccalossi <beccalossi.michele@gmail.com>